refactor: extract map runner batch summary helpers
