+++ robot_sf/benchmark/map_runner.py
@@ -14,7 +14,7 @@ from copy import deepcopy
 from dataclasses import fields
 from datetime import UTC, datetime
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, NamedTuple, TextIO
+from typing import TYPE_CHECKING, Any, TextIO
 
 import numpy as np
 import yaml
@@ -56,6 +56,18 @@ from robot_sf.benchmark.map_runner_actions import (
 )
 from robot_sf.benchmark.map_runner_actions import stack_ped_positions as _stack_ped_positions
 from robot_sf.benchmark.map_runner_actions import vel_and_acc as _vel_and_acc
+from robot_sf.benchmark.map_runner_batch_summary import (
+    WorkerMetadataBridgeUpdate as _WorkerMetadataBridgeUpdate,  # noqa: F401 - compatibility export.
+)
+from robot_sf.benchmark.map_runner_batch_summary import (
+    accumulate_batch_metadata as _accumulate_batch_metadata,  # noqa: F401 - compatibility export.
+)
+from robot_sf.benchmark.map_runner_batch_summary import (
+    apply_worker_metadata_bridge as _apply_worker_metadata_bridge,
+)
+from robot_sf.benchmark.map_runner_batch_summary import (
+    merge_runtime_algorithm_contract as _merge_runtime_algorithm_contract,
+)
 from robot_sf.benchmark.map_runner_env import (
     apply_active_observation_mode_to_env_config as _apply_active_observation_mode_to_env_config,
 )
@@ -3430,159 +3442,6 @@ def _run_map_job_worker(
         latency_stress_profile=params.get("latency_stress_profile"),
         record_simulation_step_trace=bool(params.get("record_simulation_step_trace", False)),
     )
-
-
-def _accumulate_batch_metadata(
-    rec: dict[str, Any],
-    *,
-    feasibility_totals: dict[str, float],
-) -> tuple[bool, int, int]:
-    """Aggregate adapter-impact and feasibility counters from one episode record.
-
-    Returns:
-        tuple[bool, int, int]: ``(adapter_requested_seen, native_steps, adapted_steps)`` deltas.
-    """
-    impact_meta = (rec.get("algorithm_metadata") or {}).get("adapter_impact") or {}
-    feasibility_meta = (rec.get("algorithm_metadata") or {}).get("kinematics_feasibility") or {}
-    adapter_requested_seen = False
-    adapter_native_steps = 0
-    adapter_adapted_steps = 0
-    if isinstance(impact_meta, dict):
-        adapter_requested_seen = bool(impact_meta.get("requested", False))
-        adapter_native_steps = int(impact_meta.get("native_steps", 0) or 0)
-        adapter_adapted_steps = int(impact_meta.get("adapted_steps", 0) or 0)
-    if isinstance(feasibility_meta, dict):
-        commands_evaluated = int(feasibility_meta.get("commands_evaluated", 0) or 0)
-        feasibility_totals["commands_evaluated"] += commands_evaluated
-        feasibility_totals["infeasible_native_count"] += int(
-            feasibility_meta.get("infeasible_native_count", 0) or 0
-        )
-        feasibility_totals["projected_count"] += int(
-            feasibility_meta.get("projected_count", 0) or 0
-        )
-        feasibility_totals["sum_abs_delta_linear"] += (
-            float(feasibility_meta.get("mean_abs_delta_linear", 0.0)) * commands_evaluated
-        )
-        feasibility_totals["sum_abs_delta_angular"] += (
-            float(feasibility_meta.get("mean_abs_delta_angular", 0.0)) * commands_evaluated
-        )
-        feasibility_totals["max_abs_delta_linear"] = max(
-            float(feasibility_totals["max_abs_delta_linear"]),
-            float(feasibility_meta.get("max_abs_delta_linear", 0.0) or 0.0),
-        )
-        feasibility_totals["max_abs_delta_angular"] = max(
-            float(feasibility_totals["max_abs_delta_angular"]),
-            float(feasibility_meta.get("max_abs_delta_angular", 0.0) or 0.0),
-        )
-    return adapter_requested_seen, adapter_native_steps, adapter_adapted_steps
-
-
-class _WorkerMetadataBridgeUpdate(NamedTuple):
-    """Normalized metadata update emitted by either direct or serialized worker paths."""
-
-    runtime_algorithm_contract: dict[str, Any]
-    adapter_requested_seen: bool
-    adapter_native_steps: int
-    adapter_adapted_steps: int
-
-
-def _apply_worker_metadata_bridge(
-    rec: dict[str, Any],
-    *,
-    feasibility_totals: dict[str, float],
-    runtime_algorithm_contract: dict[str, Any] | None,
-) -> _WorkerMetadataBridgeUpdate:
-    """Fold one worker episode record into the batch-level metadata contract.
-
-    This is the explicit bridge between per-episode worker payloads and the batch summary.  It is
-    used by both direct function-call execution and serialized worker execution so metadata-only
-    additions have one testable hop.
-
-    Returns:
-        Worker metadata update with merged runtime contract and adapter-impact deltas.
-    """
-    requested_seen, native_steps, adapted_steps = _accumulate_batch_metadata(
-        rec,
-        feasibility_totals=feasibility_totals,
-    )
-    merged_runtime_contract = _merge_runtime_algorithm_contract(
-        runtime_algorithm_contract or {},
-        rec.get("algorithm_metadata"),
-    )
-    return _WorkerMetadataBridgeUpdate(
-        runtime_algorithm_contract=merged_runtime_contract,
-        adapter_requested_seen=requested_seen,
-        adapter_native_steps=native_steps,
-        adapter_adapted_steps=adapted_steps,
-    )
-
-
-def _merge_runtime_algorithm_contract(  # noqa: C901
-    base_contract: dict[str, Any],
-    runtime_algorithm_metadata: Any,
-) -> dict[str, Any]:
-    """Merge runtime-resolved algorithm contract fields into a batch summary contract.
-
-    Returns:
-        dict[str, Any]: The merged contract mapping, or the original input on mismatch.
-    """
-    if not isinstance(base_contract, dict) or not isinstance(runtime_algorithm_metadata, dict):
-        return base_contract
-
-    def _merge_mapping(target: dict[str, Any], source: dict[str, Any]) -> None:
-        """Merge authoritative runtime contract values into a nested mapping."""
-        authoritative_keys = {
-            "robot_kinematics",
-            "execution_mode",
-            "adapter_name",
-            "planner_command_space",
-            "benchmark_command_space",
-            "projection_policy",
-            "execution_detail",
-            "adapter_boundary",
-        }
-
-        def _is_placeholder(value: Any) -> bool:
-            """Return whether a contract value should be replaced by runtime data."""
-            if value is None:
-                return True
-            if isinstance(value, str):
-                normalized = value.strip().lower()
-                return normalized in {"", "unknown", "unspecified", "mixed"}
-            return False
-
-        for key, value in source.items():
-            current = target.get(key)
-            if _is_placeholder(current):
-                target[key] = value
-                continue
-            if isinstance(current, dict) and isinstance(value, dict):
-                _merge_mapping(current, value)
-                continue
-            if _is_placeholder(value) or current == value:
-                continue
-            if key in authoritative_keys:
-                target[key] = value
-                continue
-            target[key] = "mixed"
-
-    runtime_planner_kinematics = runtime_algorithm_metadata.get("planner_kinematics")
-    if isinstance(runtime_planner_kinematics, dict):
-        planner_kinematics = base_contract.get("planner_kinematics")
-        if not isinstance(planner_kinematics, dict):
-            planner_kinematics = {}
-            base_contract["planner_kinematics"] = planner_kinematics
-        _merge_mapping(planner_kinematics, runtime_planner_kinematics)
-
-    runtime_upstream_reference = runtime_algorithm_metadata.get("upstream_reference")
-    if isinstance(runtime_upstream_reference, dict):
-        upstream_reference = base_contract.get("upstream_reference")
-        if not isinstance(upstream_reference, dict):
-            upstream_reference = {}
-            base_contract["upstream_reference"] = upstream_reference
-        _merge_mapping(upstream_reference, runtime_upstream_reference)
-
-    return base_contract
 
 
 def _map_result_provenance(  # noqa: PLR0913
